Schema
First crack at a latitude/longitude/point schema.
The latitude schema results in a Latitude class, whereas the longitude schema results in a primitive double.  This is just to preserve a history of 2 possible approaches.  The next commit will see one of these approaches discarded.

Adding a Gradle script for JAXB compilation.